backend/requirements.txt
fastapi
uvicorn
requests
python-multipart
fpdf2
pytest
requests-mock
httpx
